enable parallel remeshing if either metis or scotch is enabled
