+++ pytucanos/__init__.py
@@ -1,1 +1,1 @@
-from ._pytucanos import HAVE_MESHB
+from ._pytucanos import HAVE_MESHB, HAVE_PARALLEL
